Avoid gui crash due to weird engine bestmove string
When uci engine sent its weird bestmove, python-chess would raise an exception. We log this error and send a popup message to the user to change this opponent engine. The gui mode is change from Play to Neutral automatically after the user presses OK button on in the popup box.